[CH]Restructure add/edit book component
