+++ .eslintrc.json
@@ -1,3 +1,6 @@
 {
-    "extends": "airbnb"
+    "extends": "airbnb",
+    "rules": {
+        "react/jsx-filename-extension": "off"
+    }
 }
