+++ src/admin/add_book.js
@@ -1,0 +1,11 @@
+import React from 'react';
+import AddEdit from './add_edit_book';
+import './admin.css';
+
+const AddMain = () => (
+  <div className="container">
+    <AddEdit />
+  </div>
+);
+
+export default AddMain;
